+++ src/pageDocs/hour.css
@@ -1,0 +1,35 @@
+.scheduleHour {
+  background-color: white;
+  display: flex;
+  max-width: 50%;
+  flex: 1 2 200px;
+  padding: 1rem;
+  margin-top: 10px;
+  background: #ffffff 0% 0% no-repeat padding-box;
+  box-shadow: 0px 3px 6px #00000029;
+  border-radius: 5px;
+  opacity: 1;
+}
+.firstHour {
+  /* background-color: yellow; */
+  width: 60px;
+  text-align: center;
+  letter-spacing: 0;
+  color: #111111;
+  opacity: 1;
+}
+.MdChevronRight {
+  background: yellowgreen;
+}
+.iconHour {
+  /* background: bisque; */
+  align-self: center;
+  width: 60px;
+  text-align: center;
+}
+.butttonHour {
+  /* text-align: center; */
+  display: flex;
+  justify-content: flex-end;
+  outline: 1px deeppink dashed;
+}
