+++ src/pageDocs/pagDocs.css
@@ -1,16 +1,23 @@
 .containerDoc {
-  background-color: brown;
+  /* background-color: brown; */
   display: flex;
   flex-direction: row;
   justify-content: center;
   padding: 20px;
 }
 .appointment {
-  height: 821px;
-  width: 680;
-  background-color: tan;
-  width: 60vw;
-  padding: 20px;
+  /* height: 821px; */
+  /* background-color: tan; */
+  /* width: 60vw; */
+  padding: 10px;
+  margin-left: 10px;
+  /* display: flex;
+  flex-direction: column; */
+  width: 821px;
+  height: 680px;
+}
+.containerappointment {
+  color: hotpink;
 }
 .cardCardsDoc {
   background-color: white;
@@ -53,4 +60,19 @@
   color: #149bf3;
   width: 28px;
   height: 28px;
+}
+.priceDoc {
+  height: 40px;
+  background-color: #888888;
+  flex-grow: 1;
+}
+.avalibleDoc {
+  height: 40px;
+  background-color: #96d50f;
+  flex-grow: 1;
+}
+.hourDoc {
+  height: 40px;
+  background-color: #0f93d5;
+  flex-grow: 1;
 }
